requirements.txt
paho-mqtt>=1.6.1
pymodbus==2.5.3
scapy>=2.5.0
pandas>=2.0.0
scikit-learn>=1.2.0
joblib>=1.2.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
numpy>=1.24.0
pydantic>=1.10.0
pyserial-asyncio>=0.5
requests==2.32.5